chore(types): mypy-sweep Tranche 4 — qa 100->0 + pipeline 79->0; Gate = volles src/ (12 Pfade)
Letzte Tranche der GESAMTBEWERTUNG-§8-Typ-Sweeps (fresh-cache per E-057; gecachte
132/131 waren inflatiert). Verhaltensneutral: Annotations, casts, wertidentische
Wrapper, ignore-Hygiene, Overrides (+dowhy/pyod/skfolio/numba). Echte Bugs NICHT
gefixt, sondern als FIXME(mypy-sweep) + type-ignore dokumentiert:
- Zombie-Killer force-FLAT tot (_tc_signals: ops.shadow_recorder existiert nicht)
- cost_aware: OptimizerConfig als 4. Positionsarg = per_symbol_cost_bps statt
  config= (Settings still ignoriert)
- 5 tote Module in _tc_sizing (risk_budgeting/bl_sizing/mvo_optimizer/
  factor_risk_model/shadow_recorder) -> erc/bl/mvo-Sizing + Factor-Risk-Overlay
  + Shadow-Recording degradieren still zu Default/No-op
- ctx.execution_mode/ctx.equity nie gesetzt -> KPI-Write, Run-Manifest,
  Trade-Journal-Kontext, Live-Heartbeat-Site, Run-Index sterben still in excepts
- TickStore-Klasse existiert nicht (QuestDB write-through tot); OB-Imbalance-
  Merge tot (rolling_imbalance_signal liefert list[float], Caller erwartet dict)
- qa_gates: Leakage-Gate-Call mit falscher Signatur (mildernd: keine Prod-Call-
  Sites); walk_forward: toter Legacy-Pfad (nicht-existente Config-Attrs)
Neutrale Umbauten (Review-verifiziert runtime-identisch): _sizing_fn-cast-Locale
(20 Call-Sites), intel_health_flags dict[str,str]->dict[str,Any] (Annotation war
falsch), RLExecutor-Branch, _basket via TYPE_CHECKING, risk_metrics/-scenario_
engine-Rueckgabepraezisierungen (Call-Sites geprueft).
Stage-1-Findings eingearbeitet: factor_analysis float()-Wrapper in Fallback-
Pfaden revertiert (MAJOR, Verhaltensaenderung); differential_testing polars-
ignore auf [assignment, unused-ignore] (HIGH, waere CI-rot geworden — E-058).
E-058 + E-059 in docs/CLAUDE_CODING_ERRORS.md appendiert.

Verifiziert lokal: mypy --no-incremental 12 Pfade = 0 (409 Dateien; zusaetzlich
ci-debugger cold-cache unabhaengig); Tests exit 0: Zyklus (-k pipeline/
orchestrator/trading_cycle/tc_), Regression-Marker-Suite (wie CI), qa-Filter,
568 gezielte sizing/overlay/CB/intel/edcl-Tests (Stage 1); ruff clean.
CI nicht bestaetigt (Push folgt). Review-Kette: Stage 1 (risk/test/ci) PASS
nach Fixes, Stage 2 senior PASS (0 BLOCKER/MAJOR).
Follow-ups (separat, Prio): (1) Zombie-Killer-Wiring, (2) execution_mode/equity-
Wiring, (3) cost_aware-config-Arg, (4) Leakage-Gate-Signatur+Verdrahtung,
(5) OB-Imbalance, (6) Literal-Freq zentral, (7) _parse_generated_utc
AttributeError-except.

+++ src/assembled_core/qa/deflated_sharpe.py
@@ -165,7 +165,7 @@ def sharpe_threshold(
     q1 = norm.ppf(1.0 - 1.0 / n_trials)
     q2 = norm.ppf(1.0 - 1.0 / (n_trials * math.e))
     sd = math.sqrt(variance_across_trials)
-    return sd * ((1.0 - gamma) * q1 + gamma * q2)
+    return float(sd * ((1.0 - gamma) * q1 + gamma * q2))
 
 
 def deflated_sharpe(
